agregacion de accesos y reestructuracion del orden de las plantillas usadas divididas por admin,rh y supervisor

+++ asistencias/asistencias_obras/views.py
@@ -29,96 +29,17 @@ def accesos(request):
     else:
         # Puedes redirigir a una página de error o a la página de inicio, etc.
         return redirect('default_page')
-    
+
+#Funciones para administrador  
 @login_required
 def admin_dashboard(request):
     if request.user.userprofile.role != ADMIN_ROLE:
         return HttpResponseForbidden("No tienes permiso para ver esta página.")
-    return render(request, 'admin_dashboard.html')
-
-@login_required
-def rh_dashboard(request):
-    user_profile = request.user.userprofile
-    if user_profile.role == RH_ROLE:
-        # Asumiendo que 'obra_id' es un campo en el modelo de UserProfile para el ID de la obra.
-        obra_id = user_profile.obra_id
-        if obra_id:
-            obra = get_object_or_404(Obra, id=obra_id)
-            # Puedes pasar 'obra' al contexto si la plantilla necesita mostrar información sobre la obra
-            return render(request, 'rh_dashboard.html', {'obra': obra})
-        else:
-            # Manejar el caso de que no haya un ID de obra asociado
-            return HttpResponseForbidden("Este usuario de RH no tiene una obra asignada.")
-    else:
-        return HttpResponseForbidden("No tienes permiso para ver esta página.")
-
-@login_required
-def crear_empleado(request):
-    user_profile = request.user.userprofile
-    if user_profile.role == RH_ROLE:
-        obra_id = user_profile.obra_id
-        if obra_id:
-            if request.method == 'POST':
-                form = EmpleadoForm(request.POST, request.FILES)
-                if form.is_valid():
-                    empleado = form.save(commit=False)
-                    # Establece la obra al empleado antes de guardar
-                    empleado.obra_id = obra_id
-                    empleado.save()
-                    return redirect('lista_empleados')
-            else:
-                # Inicializa el formulario con la obra del usuario RH
-                form = EmpleadoForm(initial={'obra': obra_id})
-                # Oculta el campo 'obra' ya que no queremos que sea editable
-                form.fields['obra'].widget = forms.HiddenInput()
-            
-            sueldos_base = {str(puesto.id): str(puesto.sueldo_base) for puesto in Puesto.objects.all()}
-            sueldos_base_json = json.dumps(sueldos_base)
-            
-            return render(request, 'registro_empleados.html', {
-                'form': form,
-                'sueldos_base': sueldos_base_json,  # Pass 'sueldos_base' as JSON to the template
-            })
-        else:
-            return HttpResponseForbidden("Este usuario de RH no tiene una obra asignada.")
-    else:
-        return HttpResponseForbidden("No tienes permiso para ver esta página.") 
-    
-@login_required
-def lista_empleados(request):
-    user_profile = request.user.userprofile
-    if user_profile.role == RH_ROLE:
-        # Asumiendo que 'obra_id' es un campo en el modelo de UserProfile para el ID de la obra.
-        obra_id = user_profile.obra_id
-        if obra_id:
-            obra = get_object_or_404(Obra, id=obra_id)
-            # Puedes pasar 'obra' al contexto si la plantilla necesita mostrar información sobre la obra
-            empleados = Empleado.objects.filter(obra_id=obra)
-            return render(request, 'lista_empleados.html', {'empleados': empleados, 'obra': obra})
-        else:
-            # Manejar el caso de que no haya un ID de obra asociado
-            return HttpResponseForbidden("Este usuario de RH no tiene una obra asignada.")
-    else:
-        return HttpResponseForbidden("No tienes permiso para ver esta página.")
-
-
-
-@login_required
-def user_asistencia(request):
-    user_profile = request.user.userprofile
-    if user_profile.role == USER_ROLE:
-        obra_id = user_profile.obra_id
-        if obra_id:
-            obra = get_object_or_404(Obra, id=obra_id)
-            # La lógica para mostrar la asistencia del usuario en la obra
-            return render(request,'user_asistencia.html', {'obra': obra})
-        else:
-            # Manejar el caso de que no haya un ID de obra asociado
-            return HttpResponseForbidden("Este usuario no tiene una obra asignada.")
-    else:
-        return HttpResponseForbidden("No tienes permiso para ver esta página.")
+    return render(request, 'admin/admin_dashboard.html')
 
 def register(request):
+    if request.user.userprofile.role != ADMIN_ROLE:
+        return HttpResponseForbidden("No tienes permiso para ver esta página.")
     if request.method == 'POST':
         form = RegistrationForm(request.POST)
         if form.is_valid():
@@ -150,12 +71,14 @@ def register(request):
     else:
         form = RegistrationForm()
     
-    return render(request, 'registration/registro.html', {'form': form})
+    return render(request, 'admin/registro.html', {'form': form})
 
 def crear_obra(request):
     """
     Vista para crear una nueva obra. Maneja solicitudes GET para mostrar el formulario y POST para procesar el formulario.
     """
+    if request.user.userprofile.role != ADMIN_ROLE:
+        return HttpResponseForbidden("No tienes permiso para ver esta página.")
     if request.method == 'POST':
         form = ObraForm(request.POST)
         if form.is_valid():
@@ -166,26 +89,34 @@ def crear_obra(request):
         form = ObraForm()  # Inicializa un formulario en blanco para solicitudes GET
 
     # Renderiza el template con el formulario, sea nuevo o con errores
-    return render(request, 'obras/registro_obra.html', {'form': form})
+    return render(request, 'admin/registro_obra.html', {'form': form})
 
 def lista_obras(request):
     """
     Vista para listar todas las obras registradas en la base de datos.
     """
+    if request.user.userprofile.role != ADMIN_ROLE:
+        return HttpResponseForbidden("No tienes permiso para ver esta página.")
     obras = Obra.objects.all()  # Recupera todas las obras
-    return render(request, 'obras/lista_obras.html', {'obras': obras})
+    return render(request, 'admin/lista_obras.html', {'obras': obras})
 
 def cambiar_estado_obra(request, obra_id):
+    if request.user.userprofile.role != ADMIN_ROLE:
+        return HttpResponseForbidden("No tienes permiso para ver esta página.")
     obra = Obra.objects.get(id=obra_id)
     obra.activa = not obra.activa
     obra.save()
     return HttpResponseRedirect(reverse('lista_obras'))
 
 def eliminar_obra(request, obra_id):
+    if request.user.userprofile.role != ADMIN_ROLE:
+        return HttpResponseForbidden("No tienes permiso para ver esta página.")
     Obra.objects.get(id=obra_id).delete()
     return HttpResponseRedirect(reverse('lista_obras'))
 
 def editar_obra(request, obra_id):
+    if request.user.userprofile.role != ADMIN_ROLE:
+        return HttpResponseForbidden("No tienes permiso para ver esta página.")
     obra = get_object_or_404(Obra, id=obra_id)
     if request.method == 'POST':
         form = ObraForm(request.POST, instance=obra)
@@ -194,9 +125,11 @@ def editar_obra(request, obra_id):
             return redirect('lista_obras')
     else:
         form = ObraForm(instance=obra)
-    return render(request, 'obras/editar_obra.html', {'form': form, 'obra': obra})
+    return render(request, 'admin/editar_obra.html', {'form': form, 'obra': obra})
 
 def asignar_obra_a_usuario(request, user_profile_id):
+    if request.user.userprofile.role != ADMIN_ROLE:
+        return HttpResponseForbidden("No tienes permiso para ver esta página.")
     user_profile = get_object_or_404(UserProfile, pk=user_profile_id)
     if request.method == 'POST':
         form = AsignarObraForm(request.POST, instance=user_profile)
@@ -211,13 +144,82 @@ def asignar_obra_a_usuario(request, user_profile_id):
     else:
         form = AsignarObraForm(instance=user_profile)
 
-    return render(request, 'asignar_obra.html', {'form': form, 'user_profile': user_profile})
+    return render(request, 'admin/asignar_obra.html', {'form': form, 'user_profile': user_profile})
 
 def lista_user_profiles(request):
+    if request.user.userprofile.role != ADMIN_ROLE:
+        return HttpResponseForbidden("No tienes permiso para ver esta página.")
     # Si deseas filtrar por roles específicos, puedes hacerlo aquí
     user_profiles = UserProfile.objects.filter(role__in=[RH_ROLE, USER_ROLE])
-    return render(request, 'lista_user_profiles.html', {'user_profiles': user_profiles})
+    return render(request, 'admin/lista_user_profiles.html', {'user_profiles': user_profiles})
 
+
+#Funciones para RH
+@login_required
+def rh_dashboard(request):
+    user_profile = request.user.userprofile
+    if user_profile.role == RH_ROLE:
+        # Asumiendo que 'obra_id' es un campo en el modelo de UserProfile para el ID de la obra.
+        obra_id = user_profile.obra_id
+        if obra_id:
+            obra = get_object_or_404(Obra, id=obra_id)
+            # Puedes pasar 'obra' al contexto si la plantilla necesita mostrar información sobre la obra
+            return render(request, 'rh/rh_dashboard.html', {'obra': obra})
+        else:
+            # Manejar el caso de que no haya un ID de obra asociado
+            return HttpResponseForbidden("Este usuario de RH no tiene una obra asignada.")
+    else:
+        return HttpResponseForbidden("No tienes permiso para ver esta página.")
+
+@login_required
+def lista_empleados(request):
+    user_profile = request.user.userprofile
+    if user_profile.role == RH_ROLE:
+        # Asumiendo que 'obra_id' es un campo en el modelo de UserProfile para el ID de la obra.
+        obra_id = user_profile.obra_id
+        if obra_id:
+            obra = get_object_or_404(Obra, id=obra_id)
+            # Puedes pasar 'obra' al contexto si la plantilla necesita mostrar información sobre la obra
+            empleados = Empleado.objects.filter(obra_id=obra)
+            return render(request, 'rh/lista_empleados.html', {'empleados': empleados, 'obra': obra})
+        else:
+            # Manejar el caso de que no haya un ID de obra asociado
+            return HttpResponseForbidden("Este usuario de RH no tiene una obra asignada.")
+    else:
+        return HttpResponseForbidden("No tienes permiso para ver esta página.")
+
+@login_required
+def crear_empleado(request):
+    user_profile = request.user.userprofile
+    if user_profile.role == RH_ROLE:
+        obra_id = user_profile.obra_id
+        if obra_id:
+            if request.method == 'POST':
+                form = EmpleadoForm(request.POST, request.FILES)
+                if form.is_valid():
+                    empleado = form.save(commit=False)
+                    # Establece la obra al empleado antes de guardar
+                    empleado.obra_id = obra_id
+                    empleado.save()
+                    return redirect('lista_empleados')
+            else:
+                # Inicializa el formulario con la obra del usuario RH
+                form = EmpleadoForm(initial={'obra': obra_id})
+                # Oculta el campo 'obra' ya que no queremos que sea editable
+                form.fields['obra'].widget = forms.HiddenInput()
+            
+            sueldos_base = {str(puesto.id): str(puesto.sueldo_base) for puesto in Puesto.objects.all()}
+            sueldos_base_json = json.dumps(sueldos_base)
+            
+            return render(request, 'rh/registro_empleados.html', {
+                'form': form,
+                'sueldos_base': sueldos_base_json,  # Pass 'sueldos_base' as JSON to the template
+            })
+        else:
+            return HttpResponseForbidden("Este usuario de RH no tiene una obra asignada.")
+    else:
+        return HttpResponseForbidden("No tienes permiso para ver esta página.") 
+    
 @login_required
 def editar_empleado(request, empleado_id):
     empleado = get_object_or_404(Empleado, id=empleado_id)
@@ -230,4 +232,22 @@ def editar_empleado(request, empleado_id):
     else:
         form = EmpleadoForm(instance=empleado)
     
-    return render(request, 'editar_empleado.html', {'form': form, 'empleado': empleado})
+    return render(request, 'rh/editar_empleado.html', {'form': form, 'empleado': empleado})
+
+
+#Funciones para Supervisor
+@login_required
+def user_asistencia(request):
+    user_profile = request.user.userprofile
+    if user_profile.role == USER_ROLE:
+        obra_id = user_profile.obra_id
+        if obra_id:
+            obra = get_object_or_404(Obra, id=obra_id)
+            # La lógica para mostrar la asistencia del usuario en la obra
+            return render(request,'supervisor/user_asistencia.html', {'obra': obra})
+        else:
+            # Manejar el caso de que no haya un ID de obra asociado
+            return HttpResponseForbidden("Este usuario no tiene una obra asignada.")
+    else:
+        return HttpResponseForbidden("No tienes permiso para ver esta página.")
+
